src/requirements.txt
Django>=5.2,<6.0
djangorestframework>=3.14,<4.0
djangorestframework-simplejwt>=5.3,<6.0
django-cors-headers>=4.3,<5.0
python-dotenv>=1.0
Pillow>=10.0,<13.0
django-extensions>=3.2,<4.0
django-widget-tweaks>=1.5,<2.0
gunicorn>=21.2,<24.0
dj-database-url>=2.2
mysqlclient>=2.2,<3.0
psycopg2-binary>=2.9
argon2-cffi>=23.1.0,<24.0
whitenoise>=6.6.0,<7.0
